+++ server_src/constants/state.constants.js
@@ -6,6 +6,13 @@ const getFilterDefaults = () => [
   "State: Disconnected"
 ];
 
+const ERR_COUNT = {
+  offline: "offline",
+  apiKeyNotAccepted: "apiKeyNotAccepted",
+  apiKeyIsGlobal: "apiKeyIsGlobal",
+  missingSessionKey: "missingSessionKey"
+};
+
 const SYSTEM_CHECKS = {
   api: "api",
   files: "files",
@@ -114,6 +121,7 @@ const OP_STATE_FLAGS = {
 const PSTATE = {
   Offline: "Offline",
   GlobalAPIKey: "Global API Key Issue",
+  ApiKeyRejected: "API Key rejected",
   Searching: "Searching...",
   Error: "Error!",
   NoAPI: "No-API",
@@ -272,6 +280,9 @@ const mapStateToColor = (state) => {
   if (state === PSTATE.Complete) {
     return { name: "success", hex: "#00330e", category: CATEGORY.Complete };
   }
+  if (state === PSTATE.ApiKeyRejected) {
+    return { name: "danger", hex: "#2e0905", category: CATEGORY.Error };
+  }
   if (state === PSTATE.GlobalAPIKey) {
     return { name: "danger", hex: "#2e0905", category: CATEGORY.Error };
   }
@@ -291,6 +302,7 @@ module.exports = {
   getFilterDefaults,
   mapStateToColor,
   remapOctoPrintState,
+  ERR_COUNT,
   PSTATE,
   OP_STATE,
   CATEGORY,