+++ server_src/services/octoprint/octoprint-api.service.js
@@ -1,153 +1,299 @@
-const fetch = require("node-fetch");
+const fs = require("fs");
+const request = require("request");
+const {
+  OPClientErrors,
+  contentTypeHeaderKey,
+  apiKeyHeaderKey
+} = require("./constants/octoprint-service.constants");
+const { checkPluginManagerAPIDeprecation } = require("../../utils/compatibility.utils");
 const Logger = require("../../handlers/logger.js");
+const { processResponse, validatePrinter, constructHeaders } = require("./utils/api.utils");
+const { jsonContentType } = require("./constants/octoprint-service.constants");
+const { getDefaultTimeout } = require("../../constants/server-settings.constants");
 
-const logger = new Logger("OctoPrint-API-Service");
+const defaultResponseOptions = { unwrap: true };
+const octoPrintBase = "/";
+const apiBase = octoPrintBase + "api";
+const apiSettingsPart = apiBase + "/settings";
+const apiFile = (path) => apiBase + "/files/local/" + path;
+const apiFiles = (recursive = true) => apiBase + "/files?recursive=" + recursive;
+const apiConnection = apiBase + "/connection";
+const apiPrinterProfiles = apiBase + "/printerprofiles";
+const apiSystem = apiBase + "/system";
+const apiSystemInfo = apiSystem + "/info";
+const apiSystemCommands = apiSystem + "/commands";
+const apiUsers = apiBase + "/users";
+const apiLogin = apiBase + "/login?passive=true";
 
-async function fetchApi(url, method, apiKey, bodyData = undefined) {
-  return fetch(url, {
-    method,
-    headers: {
-      "Content-Type": "application/json",
-      "X-Api-Key": apiKey
-    },
-    body: JSON.stringify(bodyData)
-  });
-}
+const apiPluginManager = apiBase + "/plugin/pluginmanager";
+const apiPluginManagerRepository1_6_0 = octoPrintBase + "plugin/pluginmanager/repository";
+const apiSoftwareUpdateCheck = (force) =>
+  octoPrintBase + "plugin/softwareupdate/check" + (force ? "?force=true" : "");
+const apiPluginPiSupport = apiBase + "/plugin/pi_support";
+const apiPluginFilamentManagerSpools = apiBase + "/plugin/filamentmanager/spools";
+const apiPluginFilamentManagerProfiles = apiBase + "/plugin/filamentmanager/profiles";
+const apiTimelapse = apiBase + "/timelapse";
 
-/**
- * Fetch API with soon obsolete timeout
- * @param url
- * @param method
- * @param apiKey
- * @deprecated fetchTimeout this timeout is going to be deprecated in v1.2.0
- * @param bodyData
- * @returns {Promise<*|Promise|Promise<unknown> extends PromiseLike<infer U> ? U : (Promise|Promise<unknown>)>}
- */
-async function fetchApiTimeout(url, method, apiKey, fetchTimeout, bodyData = undefined) {
-  if (!fetchTimeout || method !== "GET" || true) {
-    return await fetchApi(url, method, apiKey, bodyData);
-  }
-  return Promise.race([
-    fetchApi(url, method, apiKey, bodyData),
-    new Promise((_, reject) =>
-      setTimeout(
-        () => reject(new Error(`timeout for url method ${method} and printer url ${url}`)),
-        fetchTimeout
-      )
-    )
-  ]);
-}
-
-class OctoPrintApiService {
+class OctoprintApiService {
   #settingsStore;
-  #timeout; // TODO this timeout is not stateful as it is derived from server settings... it might lose state
+  #httpClient;
+  #timeouts; // TODO apply apiTimeout, but apply apiRetry, apiRetryCutoff elsewhere (and webSocketRetry)
 
-  constructor({ settingsStore }) {
+  #logger = new Logger("OctoPrint-API-Service");
+
+  constructor({ settingsStore, httpClient }) {
     this.#settingsStore = settingsStore;
+    this.#httpClient = httpClient;
   }
 
-  ensureTimeoutSettingsLoaded() {
-    if (!this.#timeout) {
-      const serverSettings = this.#settingsStore.getServerSettings();
-      this.#timeout = serverSettings.timeout;
-    }
+  #ensureTimeoutSettingsLoaded() {
+    const serverSettings = this.#settingsStore.getServerSettings();
+    this.#timeouts = { ...serverSettings.timeout };
 
-    if (!this.#timeout) {
+    if (!this.#timeouts) {
       throw new Error(
         "OctoPrint API Service could not load timeout settings. settingsStore:ServerSettings:timeout didnt return anything"
       );
     }
   }
 
-  /**
-   * Retry mechanism for slow/timeout state OctoPrint entries
-   * @param printerURL
-   * @param apiKey
-   * @param item
-   * @returns {Promise<Promise<Response>|Promise<unknown> extends PromiseLike<infer U> ? U : (Promise<Response>|Promise<unknown>)|*|undefined>}
-   */
-  async getRetry(printerURL, apiKey, item) {
-    this.ensureTimeoutSettingsLoaded();
+  #prepareRequest(printer, path, timeoutOverride) {
+    this.#ensureTimeoutSettingsLoaded();
 
-    try {
-      return await this.get(printerURL, apiKey, item);
-    } catch (err) {
-      const message = `Error connecting to OctoPrint API: ${item} | ${printerURL}`;
-      logger.error(
-        `${message} | timeout: ${this.#timeout.apiTimeout}`,
-        JSON.stringify(err.message)
-      );
-      // If timeout exceeds max cut off then give up... Printer is considered offline.
-      if (this.#timeout.apiTimeout >= this.#timeout.apiRetryCutoff) {
-        logger.info(`Timeout Exceeded: ${item} | ${printerURL}`);
-        throw err;
-      }
-      // Make sure to use the settings for api retry.
-      // TODO: Fix apiRetryCutoff + apiRetry as they are swapped.
-      this.#timeout.apiTimeout = this.#timeout.apiRetryCutoff;
+    const { apiKey, printerURL } = validatePrinter(printer);
 
-      return await this.getRetry(printerURL, apiKey, item);
+    let headers = constructHeaders(apiKey);
+
+    let timeout = timeoutOverride || this.#timeouts.apiTimeout;
+    if (timeout <= 0) {
+      timeout = getDefaultTimeout().apiTimeout;
     }
+
+    return {
+      url: new URL(path, printerURL).href,
+      options: {
+        headers,
+        timeout
+      }
+    };
   }
 
-  /**
-   * Fire an action onto OctoPrint API
-   * @param printerURL
-   * @param apiKey
-   * @param route
-   * @param data
-   * @param timeout optional race to timeout (default: true)
-   * @returns {Promise<Promise<Response>|Promise<unknown> extends PromiseLike<infer U> ? U : (Promise<Response>|Promise<unknown>)>}
-   */
-  post(printerURL, apiKey, route, data, timeout = true) {
-    this.ensureTimeoutSettingsLoaded();
+  // Unused because we dont have any PUT/PATCH/POST with relevant data so far
+  #prepareJSONRequest(printer, path, data, timeoutOverride) {
+    const { url, options } = this.#prepareRequest(printer, path, timeoutOverride);
 
-    const url = new URL(route, printerURL).href;
-    return fetchApiTimeout(url, "POST", apiKey, timeout ? this.#timeout.apiTimeout : false, data);
+    // We must allow file uploads elsewhere, so be explicit about the content type and data in this JSON request
+    let serializedData = data ? JSON.stringify(data) : undefined;
+    options.headers[contentTypeHeaderKey] = jsonContentType;
+
+    return {
+      url,
+      data: serializedData,
+      options
+    };
   }
 
-  /**
-   * Acquire a GET resource
-   * @param printerURL
-   * @param apiKey
-   * @param route
-   * @param timeout optional race to timeout (default: true)
-   * @returns {Promise<Promise<Response>|Promise<unknown> extends PromiseLike<infer U> ? U : (Promise<Response>|Promise<unknown>)>}
-   */
-  get(printerURL, apiKey, route, timeout = true) {
-    this.ensureTimeoutSettingsLoaded();
+  async login(printer, responseOptions = defaultResponseOptions) {
+    const { url, options } = this.#prepareRequest(printer, apiLogin);
 
-    const url = new URL(route, printerURL).href;
-    return fetchApiTimeout(url, "GET", apiKey, timeout ? this.#timeout.apiTimeout : false);
+    const response = await this.#httpClient.post(url, {}, options);
+
+    return processResponse(response, responseOptions);
   }
 
-  /**
-   * Call a PATCH action
-   * @param printerURL
-   * @param apiKey
-   * @param route
-   * @param data body to be patched
-   * @param timeout optional race to timeout (default: true)
-   * @returns {Promise<*|Promise|Promise<unknown> extends PromiseLike<infer U> ? U : (Promise|Promise<unknown>)>}
-   */
-  patch(printerURL, apiKey, route, data, timeout = true) {
-    this.ensureTimeoutSettingsLoaded();
+  async getSettings(printer, responseOptions = defaultResponseOptions) {
+    const { url, options } = this.#prepareRequest(printer, apiSettingsPart);
 
-    const url = new URL(route, printerURL).href;
-    return fetchApiTimeout(url, "PATCH", apiKey, timeout ? this.#timeout.apiTimeout : false, data);
+    const response = await this.#httpClient.get(url, options);
+
+    return processResponse(response, responseOptions);
   }
 
-  /**
-   * Call DELETE without body data
-   * @param printerURL
-   * @param apiKey
-   * @param route
-   * @returns {Promise}
-   */
-  delete(printerURL, apiKey, route) {
-    const url = new URL(route, printerURL).href;
-    return fetchApi(url, "DELETE", apiKey);
+  async getAdminUserOrDefault(printer) {
+    const data = await this.getUsers(printer, defaultResponseOptions);
+
+    let opAdminUserName = "admin";
+    if (!!data?.users && Array.isArray(data)) {
+      const adminUser = data.users.find((user) => !!user.admin);
+      if (!adminUser) opAdminUserName = adminUser.name;
+    }
+
+    return opAdminUserName;
+  }
+
+  async getUsers(printer, responseOptions = defaultResponseOptions) {
+    const { url, options } = this.#prepareRequest(printer, apiUsers);
+
+    const response = await this.#httpClient.get(url, options);
+
+    return processResponse(response, responseOptions);
+  }
+
+  async getFiles(printer, recursive = false, responseOptions = defaultResponseOptions) {
+    const { url, options } = this.#prepareRequest(printer, apiFiles(recursive));
+
+    const response = await this.#httpClient.get(url, options);
+
+    return processResponse(response, responseOptions);
+  }
+
+  async getFile(printer, path, responseOptions = defaultResponseOptions) {
+    const { url, options } = this.#prepareRequest(printer, apiFile(path));
+
+    const response = await this.#httpClient.get(url, options);
+
+    return processResponse(response, responseOptions);
+  }
+
+  async getConnection(printer, responseOptions = defaultResponseOptions) {
+    const { url, options } = this.#prepareRequest(printer, apiConnection);
+
+    const response = await this.#httpClient.get(url, options);
+
+    return processResponse(response, responseOptions);
+  }
+
+  async getPrinterProfiles(printer, responseOptions = defaultResponseOptions) {
+    const { url, options } = this.#prepareRequest(printer, apiPrinterProfiles);
+
+    const response = await this.#httpClient.get(url, options);
+
+    return processResponse(response, responseOptions);
+  }
+
+  async getPluginManager(printer, responseOptions = defaultResponseOptions) {
+    const printerManagerApiCompatible = checkPluginManagerAPIDeprecation(printer.octoPrintVersion);
+
+    const path = printerManagerApiCompatible ? apiPluginManagerRepository1_6_0 : apiPluginManager;
+    const { url, options } = this.#prepareRequest(printer, path);
+
+    const response = await this.#httpClient.get(url, options);
+
+    return processResponse(response, responseOptions);
+  }
+
+  async getSystemInfo(printer, responseOptions = defaultResponseOptions) {
+    const { url, options } = this.#prepareRequest(printer, apiSystemInfo);
+
+    const response = await this.#httpClient.get(url, options);
+
+    return processResponse(response, responseOptions);
+  }
+
+  async getSystemCommands(printer, responseOptions = defaultResponseOptions) {
+    const { url, options } = this.#prepareRequest(printer, apiSystemCommands);
+
+    const response = await this.#httpClient.get(url, options);
+
+    return processResponse(response, responseOptions);
+  }
+
+  async getSoftwareUpdateCheck(printer, force, responseOptions = defaultResponseOptions) {
+    const { url, options } = this.#prepareRequest(printer, apiSoftwareUpdateCheck(force));
+
+    const response = await this.#httpClient.get(url, options);
+
+    return processResponse(response, responseOptions);
+  }
+
+  async getPluginPiSupport(printer, responseOptions = defaultResponseOptions) {
+    const { url, options } = this.#prepareRequest(printer, apiPluginPiSupport);
+
+    const response = await this.#httpClient.get(url, options);
+
+    return processResponse(response, responseOptions);
+  }
+
+  async deleteTimeLapse(printer, fileName, responseOptions = defaultResponseOptions) {
+    if (!fileName) {
+      throw new Error("Cant delete timelapse file without providing filename");
+    }
+
+    const path = `${apiTimelapse}/${fileName}`;
+    const { url, options } = this.#prepareRequest(printer, path);
+
+    const response = await this.#httpClient.delete(url, options);
+
+    return processResponse(response, responseOptions);
+  }
+
+  async listUnrenderedTimeLapses(printer, responseOptions = defaultResponseOptions) {
+    const path = `${apiTimelapse}?unrendered=true`;
+    const { url, options } = this.#prepareRequest(printer, path);
+
+    const response = await this.#httpClient.get(url, options);
+
+    return processResponse(response, responseOptions);
+  }
+
+  async listPluginFilamentManagerProfiles(printer, responseOptions = defaultResponseOptions) {
+    const { url, options } = this.#prepareRequest(printer, apiPluginFilamentManagerProfiles);
+
+    const response = await this.#httpClient.get(url, options);
+
+    return processResponse(response, responseOptions);
+  }
+
+  async listPluginFilamentManagerFilament(printer, responseOptions = defaultResponseOptions) {
+    const { url, options } = this.#prepareRequest(printer, apiPluginFilamentManagerSpools);
+
+    const response = await this.#httpClient.get(url, options);
+
+    return processResponse(response, responseOptions);
+  }
+
+  async getPluginFilamentManagerFilament(
+    printer,
+    filamentID,
+    responseOptions = defaultResponseOptions
+  ) {
+    // filamentID needs to be INT numeric
+    // https://github.com/malnvenshorn/OctoPrint-FilamentManager/blob/647af691d6081df2f16d400e834f12f11f6eea56/octoprint_filamentmanager/data/__init__.py#L84
+    const parsedFilamentID = Number.parseFloat(filamentID);
+    if (isNaN(filamentID)) {
+      throw OPClientErrors.filamentIDNotANumber;
+    }
+
+    const path = `${apiPluginFilamentManagerSpools}/${parsedFilamentID}`;
+    const { url, options } = this.#prepareRequest(printer, path);
+
+    const response = await this.#httpClient.get(url, options);
+
+    return processResponse(response, responseOptions);
+  }
+
+  // TODO WIP with axios
+  async downloadFile(printerConnection, fetchPath, targetPath, callback) {
+    const fileStream = fs.createWriteStream(targetPath);
+
+    // https://stackoverflow.com/questions/55374755/node-js-axios-download-file-stream-and-writefile
+
+    // TODO
+    const res = await this.getWithOptionalRetry(printerConnection, fetchPath, false);
+
+    return await new Promise((resolve, reject) => {
+      res.body.pipe(fileStream);
+      res.body.on("error", reject);
+      fileStream.on("finish", async () => {
+        await callback(resolve, reject);
+      });
+    });
+  }
+
+  // TODO WIP
+  async downloadImage({ printerURL, apiKey }, fetchPath, targetPath, callback) {
+    const fileStream = fs.createWriteStream(targetPath);
+
+    // https://stackoverflow.com/questions/55374755/node-js-axios-download-file-stream-and-writefile
+
+    // TODO
+    const downloadURL = new URL(fetchPath, printerURL);
+    return request.head(downloadURL, (err, res, body) => {
+      res.headers[contentTypeHeaderKey] = "image/png";
+      res.headers[apiKeyHeaderKey] = apiKey;
+      request(url).pipe(fs.createWriteStream(targetPath)).on("close", callback);
+    });
   }
 }
 
-module.exports = OctoPrintApiService;
+module.exports = OctoprintApiService;